+++ gen_docs.sh
@@ -5,7 +5,7 @@ set -e
 
 branch="$(git rev-parse --abbrev-ref HEAD)"
 checker=$(git diff --name-only `git merge-base $branch HEAD`)
-if [[ ! $checker =~ "version.py" ]]; then
+if [[ ! $checker =~ "__init__.py" ]]; then
   echo -e "\n********************************************************************ERROR********************************************************************"
   echo "Docs generation was ABORTED since module version was not bumped!! Changelog generator requires the commit number and package version in sync."
   echo -e "*********************************************************************************************************************************************\n"
@@ -19,7 +19,8 @@ if [[ ! $checker =~ "release_notes.rst" ]]; then
   exit 255
 fi
 
-rm -rf docs && mkdir docs
-[ ! -d "doc_generator/_static" ] && mkdir doc_generator/_static  # creates a _static folder if unavailable
+rm -rf docs
+mkdir docs
+mkdir -p doc_generator/_static  # creates a _static folder if unavailable
 cp README.md doc_generator && cd doc_generator && make clean html && mv _build/html/* ../docs && rm README.md
 touch ../docs/.nojekyll