+++ gmailconnector/__init__.py
@@ -1,0 +1,8 @@
+"""Place holder for package."""
+
+from gmailconnector.options import Category, Condition, Folder  # noqa: F401
+from gmailconnector.read_email import ReadEmail  # noqa: F401
+from gmailconnector.send_email import SendEmail  # noqa: F401
+from gmailconnector.send_sms import SendSMS  # noqa: F401
+
+version = "0.6.6"